+++ internal/ingest/handler.go
@@ -52,7 +52,7 @@ func (h *ingestHandler) ServeTrack(tw *moqt.TrackWriter) {
 	case trackCatalog:
 		src = h.catalog
 	default:
-		tw.CloseWithError(moqt.TrackNotFoundErrorCode)
+		tw.CloseWithError(moqt.SubscribeErrorCodeNotFound)
 		return
 	}
 	src.serve(tw, h.done)
@@ -233,7 +233,7 @@ func (s *trackSource) serve(tw *moqt.TrackWriter, done <-chan struct{}) {
 				f := g.next(frameIdx)
 				if f != nil {
 					if err := gw.WriteFrame(f); err != nil {
-						gw.Close()
+						_ = gw.Close()
 						return
 					}
 					frameIdx++
@@ -249,15 +249,15 @@ func (s *trackSource) serve(tw *moqt.TrackWriter, done <-chan struct{}) {
 				case <-notify:
 				case <-time.After(notifyTimeout):
 				case <-done:
-					gw.Close()
+					_ = gw.Close()
 					return
 				case <-twCtx.Done():
-					gw.Close()
+					_ = gw.Close()
 					return
 				}
 			}
 
-			gw.Close()
+			_ = gw.Close()
 			continue
 		}